Begin implementing @intrinsic functions.
introducing @intrinsinc funcs

I have some doubts at this point.

Why not give as argument the name of the intrinsic, in case
the platform name cannot be expressed as a symbol in ion?

Can intrinsic be renamed in symbols?

 a := intrinsic_func
 a();

It seems this should not be accepted.

And what about this:

 a = another_intrinsic_func;
 a();
?

An always inline foreign function vs an intrinsic. Are they the same?
Should intrinsic share types with func?