Fix issue with stdout constant on windows, don't define @foreign vars.
By not defining @foreign vars, we fix an issue whereas we would not be
able to define stdout ourselves, as it would conflict with the pre-existing
macros.

If this patch is too broad, then we need to find a way to control more
finely what gets defined and what does not get defined.

Generally the issue with not defining things ourselves, is that we may
not be able to detect discrepancies between our definition types and the
foreign definition types.

The problem with defining things ourselves, is that we must accurately
describe what they are.

stdin, stdout and stderr are described in C as being provided as macros,
as expressions, not constants. For the purpose of our interface, maybe
we should describe them as such.