Add `@printf(fmt)` note for printf sanity checks.
With clang and gcc, an attribute is added, which will allow the C code
to sanity check the format string and the function's arguments.

I've tried enabling the same annotation for MSVC (Sal annotation)
without luck, since my test example came out with an out of memory error
when I tried `cl /analyze` on it.